refactor :: remove contract
