+++ presentation/src/main/kotlin/team/retum/jobis_android/feature/auth/resetpassword/ResetPasswordViewModel.kt
@@ -1,5 +1,6 @@
 package team.retum.jobis_android.feature.auth.resetpassword
 
+import androidx.annotation.StringRes
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
@@ -22,6 +23,8 @@ import team.retum.domain.usecase.user.SendVerificationCodeUseCase
 import team.retum.domain.usecase.user.VerifyEmailUseCase
 import team.retum.jobis_android.feature.root.BaseViewModel
 import team.retum.jobis_android.util.Regex
+import team.retum.jobis_android.util.mvi.SideEffect
+import team.retum.jobis_android.util.mvi.State
 import javax.inject.Inject
 
 @HiltViewModel
@@ -245,4 +248,28 @@ internal class ResetPasswordViewModel @Inject constructor(
             )
         }
     }
+}
+
+data class ResetPasswordState(
+    val email: String = "",
+    val authCode: String = "",
+    val currentPassword: String = "",
+    val newPassword: String = "",
+    val passwordRepeat: String = "",
+    val emailErrorState: Boolean = false,
+    val authCodeErrorState: Boolean = false,
+    val sendAuthCodeState: Boolean = false,
+    val passwordFormatErrorState: Boolean = false,
+    val passwordRepeatErrorState: Boolean = false,
+    val comparePasswordErrorState: Boolean = false,
+    val buttonEnabled: Boolean = false,
+) : State
+
+sealed class ResetPasswordSideEffect : SideEffect {
+    object SuccessVerification : ResetPasswordSideEffect()
+    object SuccessChangePassword : ResetPasswordSideEffect()
+    object SuccessResetPassword : ResetPasswordSideEffect()
+    object PasswordMismatch : ResetPasswordSideEffect()
+    object ClearFocus : ResetPasswordSideEffect()
+    class Exception(@StringRes val message: Int) : ResetPasswordSideEffect()
 }
