--- presentation/src/main/kotlin/team/retum/jobis_android/feature/auth/signin/SignInContract.kt
@@ -1,19 +1,0 @@
-package team.retum.jobis_android.feature.auth.signin
-
-import androidx.annotation.StringRes
-import team.retum.jobis_android.util.mvi.SideEffect
-import team.retum.jobis_android.util.mvi.State
-
-data class SignInState(
-    val email: String = "",
-    val password: String = "",
-    val autoSignIn: Boolean = false,
-    val emailError: Boolean = false,
-    val passwordError: Boolean = false,
-    val signInButtonEnabled: Boolean = false,
-) : State
-
-sealed class SignInSideEffect : SideEffect {
-    object MoveToMain : SignInSideEffect()
-    class Exception(@StringRes val message: Int) : SignInSideEffect()
-}
