+++ presentation/src/main/kotlin/team/retum/jobis_android/feature/auth/signup/SignUpViewModel.kt
@@ -1,5 +1,6 @@
 package team.retum.jobis_android.feature.auth.signup
 
+import androidx.annotation.StringRes
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
@@ -22,6 +23,8 @@ import team.retum.domain.usecase.user.SignUpUseCase
 import team.retum.domain.usecase.user.VerifyEmailUseCase
 import team.retum.jobis_android.feature.root.BaseViewModel
 import team.retum.jobis_android.util.Regex
+import team.retum.jobis_android.util.mvi.SideEffect
+import team.retum.jobis_android.util.mvi.State
 import java.util.regex.Pattern
 import javax.inject.Inject
 
@@ -357,4 +360,44 @@ class SignUpViewModel @Inject constructor(
         `class`: String,
         number: String,
     ) = Integer.parseInt("$grade$`class`${number.padStart(2, '0')}")
+}
+
+data class SignUpState(
+    val email: String = "",
+    val emailError: Boolean = false,
+    val verifyCode: String = "",
+    val verifyCodeError: Boolean = false,
+    val password: String = "",
+    val passwordError: Boolean = false,
+    val repeatPassword: String = "",
+    val repeatPasswordError: Boolean = false,
+    val grade: String = "",
+    val name: String = "",
+    val gender: Gender = Gender.MAN,
+    val classRoom: String = "",
+    val number: String = "",
+    val studentNotFound: Boolean = false,
+    val sendVerifyCodeButtonEnabled: Boolean = false,
+    val authCodeEnabled: Boolean = false,
+    val signUpButtonEnabled: Boolean = false,
+) : State
+
+sealed class SignUpSideEffect : SideEffect {
+    sealed class StudentInfo {
+        object CheckStudentExistsSuccess : SignUpSideEffect()
+        object CheckStudentExistsNotFound : SignUpSideEffect()
+    }
+
+    sealed class VerifyEmail {
+        object EmailConflict : SignUpSideEffect()
+        object VerifyEmailSuccess : SignUpSideEffect()
+        object SendAuthCodeSuccess : SignUpSideEffect()
+    }
+
+    sealed class SetPassword {
+        object SignUpSuccess : SignUpSideEffect()
+        object SignUpConflict : SignUpSideEffect()
+    }
+
+    class Exception(@StringRes val message: Int) : SignUpSideEffect()
 }
